+++ backend/database/connectDataBase.js
@@ -5,8 +5,8 @@ const sequelize = new Sequelize(process.env.SEQUELIZE_URL, { logging: false });
 const connectDataBase = () => {
     sequelize
         .authenticate()
-        .then((data) => console.log('Connecting with db successful...'))
-        .catch((err) => {
+        .then(data => console.log('Connecting with db successful...'))
+        .catch(err => {
             console.log('Connecting ends with error...');
         });
 };
