+++ backend/database/tables/storage.js
@@ -1,0 +1,26 @@
+const { sequelize, DataTypes } = require('../connectDataBase');
+
+const Storage = sequelize.define(
+    'storage',
+    {
+        storage_id: {
+            type: DataTypes.INTEGER,
+            primaryKey: true,
+            autoIncrement: true,
+        },
+        name: {
+            type: DataTypes.STRING,
+            allowNull: false,
+        },
+        is_deleted: {
+            type: DataTypes.BOOLEAN,
+            defaultValue: false,
+            allowNull: false,
+        },
+    },
+    {
+        freezeTableName: true,
+    }
+);
+
+module.exports.Storage = Storage;
