+++ backend/server.js
@@ -6,7 +6,7 @@ const { connectDataBase } = require('../backend/database/connectDataBase');
 const {
     addNewUser,
     returnUsers,
-    deleteUser,
+    isDeleteUser,
 } = require('../backend/database/editingUser');
 
 connectDataBase();
@@ -14,20 +14,21 @@ connectDataBase();
 app.use(express.json());
 app.use(cors());
 
-
 app.post('/api/user', async (req, res) => {
-    await addNewUser(req.body);
+    const a = await addNewUser(req.body);
+    res.status(201).send(a)
+    
 });
 
 app.delete('/api/user', async (req, res) => {
-    await deleteUser(req.body);
+    const d = await isDeleteUser(req.body);
+    res.status(200).send(d)
 });
 
 app.get('/api/user', async (req, res) => {
     const users = await returnUsers();
-    await res.json(JSON.parse(users));
+    await res.send(users);
 });
-
 
 app.listen(process.env.PORT, (error) => {
     error
