+++ frontend/viewes/index.html
@@ -42,7 +42,9 @@
         </div>
 
         <div class="container py-3">
-            <div class="showUser row"></div>
+            <div class="showUser row">
+                
+            </div>
         </div>
 
         <script src="../js/code.jquery.com_jquery-3.7.0.min.js"></script>
