+++ package.json
@@ -19,12 +19,14 @@
   "homepage": "https://github.com/yaGlist/WareHouse#readme",
   "devDependencies": {
     "dotenv": "^16.3.1",
+    "eslint": "^8.44.0",
     "nodemon": "^2.0.22"
   },
   "dependencies": {
     "cors": "^2.8.5",
     "ejs": "^3.1.9",
     "express": "^4.18.2",
+    "loadash": "^1.0.0",
     "pg": "^8.11.1",
     "pg-hstore": "^2.3.4",
     "sequelize": "^6.32.1"
